♻️Refactor(*): 프로젝트 폴더 구조 변경
- 도메인별로 엔드포인트를 라우터에 등록해서 관리
- main에서는 각 엔드포인트에 대한 라우터를 import
