+++ BE/app/domain/messenger/router/__init__.py
@@ -1,0 +1,5 @@
+# from fastapi import APIRouter
+# from . import websocket_controller
+
+# router = APIRouter()
+# router.include_router(websocket_controller.router, prefix="/ws", tags=["websocket"])
